Snapshot testing for schemas.
