+++ graphql-parser/tests/schemas/fail_onion.graphql
@@ -1,0 +1,1 @@
+onion Feed = Story | Article | Advert
